Feat(main_page.dart): 메인 페이지, 검색 페이지, 설정 페이지 init
