+++ lib/pages/search_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class SearchPage extends StatefulWidget {
+  const SearchPage({super.key});
+
+  @override
+  State<SearchPage> createState() => _SearchPageState();
+}
+
+class _SearchPageState extends State<SearchPage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
